Implement random placement selection
Since this is going to be called in a hot loop, I spent some time
benchmarking different options. The implementations I tried were:

- Iterate once calling `rng.gen_weighted_bool(i+1)` and grabbing the
  last element
- Iterate an extra time to count and then call `nth`
- Iterate once, collect into Vec

The vec option ended up being most efficient (mostly because of the
allocation in the call to `valid_placements`). If we ever get closures
which can implement clone, the second option is probably just as
efficient.