Actually implement real MCTS
This makes one major change to what we were doing before for naive
implementations, which is that we keep track of the total DPS we've
seen and normalize that based on the highest DPS we've ever seen. By
doing that, we end up with a score per exploration that is bounded
between 0 and 1, and brings us to the classical multi-armed bandit
problem.

To test this out, I made two changes to the data. The first was to put
in the stats for Sally's gear, so that she was obviously the best DPS
and so that it would have the conditional buff from 4 crusaders being
adjacent. I also added a 10th crusader, with no stats and 0 DPS -- With
only 9 crusaders to fill 10 slots it would try to place Sally in (1, 1)
which meant that Veronica could never be placed without being adjacent.
This is because if you're picking random formations, and one of the
slots will always be empty, you are more likely to end up with a good
answer if Sally is in the one spot where she could never have less than
4 adjacent to her.

Testing seemed quite promising. Most runs it would quickly realize that
the best spots for Sally are (1, 0) and (1, 2), which is where she can
fulfill all her buffs, Grandmora can be optimally placed, and Veronica
can buff her without being adjacent. Interestingly, it'd trend toward
Grandmora being in (0, 1) and (0, 2) -- Presumably so she leaves the
corners open for Larry. I'll be interested to see the trends once all
the data is in place.

Right now this is actually super inefficient, since we're re-calculating
the formation's DPS multiple times per pass. We can optimize that later
if we feel the need to, but we're actually getting reasonable answers
without doing that.