Implement slot 9 basic abilities
I've also switched to highest DPS and then most checked in terms of
choosing which branch we take. They're usually the same in early stages,
but when we get close to the end where we will have fully expanded the
tree, the old version would always favor choosing a slot with fewer
options, as that leads to more potential nodes to check.