Fix over-convergence of search
When a node became fully expanded, it would always add its highest DPS
to the total seen rather than the DPS of the formation it was trying.
This meant that once any node became fully expanded, we would end up
heavily biased towards continuing to try it.

The result of this was that sometimes we'd end up fully expanding a
sub-optimal node (e.g. placing Sally at (1, 1)) because we happened to
see one good formation for it, causing the search to focus purely on
that node and become unlikely to find that placing Sally at (1, 2) would
lead to a higher DPS

This change now more correctly distributes its search. This means we'll
have a lower confidence factor in the move we do end up making, but we
should more consistently pick the best placement.