Implement ability buffs
Several crusaders have abilities that simply buff other abilities
(sometimes conditionally). THis adds the basic framework needed for
those cases. This should work fine for gear as well. It won't yet handle
things like "Add N more targets" or anything like that.

I've implemented in such a way that we should be able to easily move to
a map lookup by tag if we find that it'll be a significant performance
improvement.