Replace `Lt` and `Gt` conditions with `LtEq` and `GtEq`
I had written Snickette's condition for focused support incorrectly, and
as a result she was basically never being considered. I typically try
to write things so that the numbers are exactly what the game displays
(this is why, for example, percentages are 100.0 and not 1.0) in order
to avoid these kinds of bugs.

I also had misread Sisaron's buff, and have corrected that case as well.