Propogate search results to sibling nodes
We have a problem of duplicate information. For our purposes, 1-2-3 and
1-3-2 are the same, but are tracked as different nodes. In an earlier
implementation I had tried to actually solve this by having the children
stored as `Rc<RefCell<Child>>`. While that potentially helps with memory
pressure, it ultimately doesn't achieve anything that this
implementation does not, but requires much more complicated code.