Allow search configuration options to be passed via CLI
This is limited purely to parameters which were already configurable.
That means it doesn't handle the two most important cases, which
formation to use, and user data. However, for the options that we can
handle easily, we shouldn't require recompiling code to change how long
it searches or how much gold to use.