Allow configuration of active play abilities from CLI
This was mostly motivated by setting the groundwork to consider gold,
but it ended up being very easy to quickly add an option to include or
exclude abilities which require active play. I'm not sure if there are
any other implemented abilities which should have this flag, but these
are the obvious cases.