Allow user data to be loaded from a TOML file
This file is a pain in the ass to fill, and I'd eventually like to
provide a save import utility. However, that'll be a bit more difficult
to write, and at least this file is meant to be edited by humans and
doesn't require recompiling code.

Talents don't get added very often, so I'm strict about input there.
However, crusaders are added every 2 weeks, and there's already a ton I
don't support (I've even unlocked some since starting work on this), so
for unrecognized crusaders I just go ahead and print a warning. Once
we're more-or-less caught up, I may want to make this more strict.