Fix a few numerical bugs
Testing a few formations I've been using to make sure the numbers match
up exactly. I had messed up Billy's legendary (the wiki lied), and
shockingly Artaxes' effect is multiplicative not additive, even though
Sasha has an identical effect that is additive. I'll have to keep an eye
on that number to see if it changes