Consider gold-find %
I'm treating gold find simply as a straight % buff in terms of score,
since doubling DPS means monsters die twice as quickly and therefore
give twice the gold per second, same as if you were to double your gold
find. This is only true if you're not insta-killing, but neither gold
nor DPS actually really matter if you're still insta-killing.

I've enabled gold-find consideration by default, with an option to
disable it if you want to rank on DPS only. I've verified that the
number are correct at least for the formation that I was using when I
wrote this. I did not know that Larry's little pockets was
multiplicative.

Since the effect is also affected by gear, I had to swap the order in
which the two modifiers were applied. This required a bit of additional
legwork, since I've written effects that are in the form of "x% + y% for
each z" assuming that the gear multpilier happens after the plus/minus.
To fix this, I now make sure that we check for gear when we're doing a
plus/minus effect both on the main % and also on the % that we're
adding/subtracting.

For some reason after this change, some limited testing is landing on
much better formations than it was before. It's possible that it's just
a fluke -- But I think it may be due to Gold Panda always getting placed
early, causing Sally to have a female on the field to influence her
placement. It's also now less biased towards putting her (or any primary
DPS) in the center, since it's now *heavily* biased towards placing
Larry somewhere with 6 adjacent crusaders.