Cache search results at each step
This changes the characteristics of the program quite a bit if it's run
multiple times. Rather than starting from scratch each time we do a
search, we instead save the results to disk. Each time we are about to
start a new iteration, we see if we have any information about that
state, and load from disk if so.

This means that subsequent runs of the program will refine the previous
search rather than starting from scratch. Each iteration is stored
completely independent. So loading the root search from cache won't use
the information it gained after taking the first step. This is because
doing so would strongly bias the search towards whatever path it took
last time. This may be something that we can overcome in the future,
possibly by choosing the next node based on average DPS rather than
times visited.